ml/models/baseline/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
pyyaml>=6.0
shap>=0.44.0
matplotlib>=3.8.0
